Feature/access control

+++ src/routes/plants.js
@@ -8,6 +8,7 @@ import {
   updatePlant
 } from "../db/plants.js";
 import { protect } from "../middleware/authMiddleware.js";
+import { adminOnly } from "../middleware/adminMiddleware.js";
 
 const router = Router();
 
@@ -17,6 +18,24 @@ router.get('/', async (req, res) => {
     res.json(plants);
   } catch (error) {
     res.status(500).json({ error: "Failed to fetch plants" });
+  }
+});
+
+router.get("/mine", protect, async (req, res) => {
+  try {
+    const plants = await getPlantsByOwner(req.user.userId);
+    res.json(plants);
+  } catch (error) {
+    res.status(500).json({ error: "Failed to fetch your plants" });
+  }
+});
+
+router.get('/owner/:ownerId', protect, adminOnly, async (req, res) => {
+  try {
+    const plants = await getPlantsByOwner(req.params.ownerId);
+    res.json(plants);
+  } catch (error) {
+    res.status(500).json({ error: "Failed to fetch plants by owner" });
   }
 });
 
@@ -33,6 +52,7 @@ router.get('/:id', async (req, res) => {
 });
 
 router.post('/', protect, async (req, res) => {
+  console.log("REQ USER IN POST:", req.user);
   try {
     const newPlant = await createPlant({
       ...req.body,
@@ -47,22 +67,24 @@ router.post('/', protect, async (req, res) => {
 
 router.delete('/:id', protect, async (req, res) => {
   try {
-    const deletedPlant = await deletePlant(req.params.id);
-    if (!deletedPlant) {
+    const plant = await getPlantById(req.params.id);
+
+    if (!plant) {
       return res.status(404).json({ error: "Plant not found" });
     }
+
+    if (
+      plant.owner.toString() !== req.user.userId &&
+      req.user.role !== "admin"
+    ) {
+      return res.status(403).json({ error: "Not allowed" });
+    }
+
+    await deletePlant(req.params.id);
+
     res.json({ message: "Plant deleted successfully" });
   } catch (error) {
     res.status(500).json({ error: "Failed to delete plant" });
-  }
-});
-
-router.get('/owner/:ownerId', async (req, res) => {
-  try {
-    const plants = await getPlantsByOwner(req.params.ownerId);
-    res.json(plants);
-  } catch (error) {
-    res.status(500).json({ error: "Failed to fetch plants by owner" });
   }
 });
 
@@ -87,6 +109,12 @@ router.put('/:id', protect, async (req, res) => {
   } catch (error) {
     res.status(500).json({ error: "Failed to update plant" });
   }
+});
+
+
+router.delete("/admin/:id", protect, adminOnly, async (req, res) => {
+  await deletePlant(req.params.id);
+  res.json({ message: "Deleted by admin" });
 });
 
 export default router;
